@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "kube-copilot"
-version = "0.1.23"
+version = "0.1.24"
 description = "Kubernetes Copilot"
 authors = ["Pengfei Ni <feiskyer@gmail.com>"]
 readme = "README.md"
@@ -13,18 +13,21 @@ keywords = ["kubernetes", "copilot", "openai", "chatgpt"]
 kube-copilot = 'kube_copilot.cli:main'
 
 [tool.poetry.dependencies]
-python = ">=3.10,<4"
+python = ">=3.11,<4"
 langchain = ">=0.0.180"
 requests = ">=2.31"
-openai = ">=0,<1"
+openai = ">1"
 tiktoken = ">=0.3.1"
 lark = "==1.1.8"
 faiss-cpu = {version = "^1"}
 google-api-python-client = ">=2.85.0"
 click = ">=8.1.3"
-kubernetes = ">=26.1,<28.0"
+kubernetes = ">=28.0"
 streamlit = "^1.24.0"
-langchain-experimental = ">=0.0.13,<0.0.38"
+langchain-experimental = ">0.0.38"
+
+[tool.poetry.group.dev.dependencies]
+httpx = {extras = ["socks"], version = "^0.26.0"}
 
 [build-system]
 requires = ["poetry-core"]
